# extracted_project/Session-Guard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram Broadcast Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bots/{botId}/setup-webhook:
    parameters:
      - $ref: "#/components/parameters/BotId"
    post:
      operationId: setupWebhook
      tags: [bots]
      summary: Register Telegram webhook for a bot
      responses:
        "200":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSetupResponse"
...
components:
  parameters:
...
    BotId:
      name: botId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    WebhookSetupResponse:
      type: object
      properties:
        webhookUrl:
          type: string
      required: [webhookUrl]
...